Adding an instrumentation test.
In an effort to try and find some performance bottlenecks, I'm adding some dynamic instrumentation of the Cuid2.cfc. This will look at the aggregate execution times of the various methods.

One thing I'm shocked to see is that the `secureRandRange()` method accounts for something like 30% of the overall execution times.